+++ KXMapStudio.App/Services/Pack/IPackLoader.cs
@@ -1,0 +1,9 @@
+using KXMapStudio.Core;
+using System.Threading.Tasks;
+
+namespace KXMapStudio.App.Services.Pack;
+
+public interface IPackLoader
+{
+    Task<PackLoadResult> LoadPackAsync(string path);
+}
