+++ KXMapStudio.App/Services/Pack/SingleFilePackLoader.cs
@@ -1,0 +1,18 @@
+using KXMapStudio.Core;
+using System.Collections.Generic;
+using System.IO;
+using System.Threading.Tasks;
+
+namespace KXMapStudio.App.Services.Pack;
+
+public class SingleFilePackLoader : IPackLoader
+{
+    public async Task<PackLoadResult> LoadPackAsync(string path)
+    {
+        var originalRawContent = new Dictionary<string, byte[]>(StringComparer.OrdinalIgnoreCase);
+        originalRawContent[Path.GetFileName(path)] = await File.ReadAllBytesAsync(path);
+
+        var packLoader = new PackLoader();
+        return await packLoader.LoadPackFromMemoryAsync(originalRawContent, path, false);
+    }
+}
