+++ KXMapStudio.App/State/WorkspaceManager.cs
@@ -1,0 +1,169 @@
+using KXMapStudio.App.Services;
+using KXMapStudio.Core;
+using Microsoft.Extensions.Logging;
+using Microsoft.Win32;
+using System.IO;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Xml.Linq;
+
+namespace KXMapStudio.App.State;
+
+public class WorkspaceManager
+{
+    private readonly PackService _packService;
+    private readonly PackWriterService _packWriterService;
+    private readonly IDialogService _dialogService;
+    private readonly ILogger<WorkspaceManager> _logger;
+
+    public WorkspaceManager(
+        PackService packService,
+        PackWriterService packWriterService,
+        IDialogService dialogService,
+        ILogger<WorkspaceManager> logger)
+    {
+        _packService = packService;
+        _packWriterService = packWriterService;
+        _dialogService = dialogService;
+        _logger = logger;
+    }
+
+    public async Task<(LoadedMarkerPack? pack, string? path)> OpenWorkspaceAsync(string path)
+    {
+        try
+        {
+            var result = await _packService.LoadPackAsync(path);
+            if (result.LoadedPack == null || !result.LoadedPack.XmlDocuments.Any())
+            {
+                _logger.LogWarning("PackService returned no XML documents for path: {Path}", path);
+                _dialogService.ShowError("No XML Files Found", "The selected folder or pack does not contain any valid .xml marker files.");
+                return (null, null);
+            }
+
+            if (result.HasErrors)
+            {
+                ShowLoadPackErrors(result);
+            }
+
+            return (result.LoadedPack, path);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, "A catastrophic error occurred while loading from {Path}", path);
+            _dialogService.ShowError("An Unexpected Error Occurred", $"A critical error occurred while trying to open the workspace.\n\nDetails: {ex.Message}");
+            return (null, null);
+        }
+    }
+
+    public async Task SaveActiveDocumentAsync(LoadedMarkerPack workspacePack, string activeDocumentPath, string workspacePath)
+    {
+        if (activeDocumentPath.StartsWith("Untitled"))
+        {
+            await SaveActiveDocumentAsAsync(workspacePack, activeDocumentPath);
+            return;
+        }
+
+        if (workspacePath == null)
+        {
+            _logger.LogError("Attempted to save an existing document, but WorkspacePath is null.");
+            _dialogService.ShowError("Save Error", "Cannot save the file because the original path is unknown. Please use 'Save As...'.");
+            return;
+        }
+
+        string fullSavePath = Directory.Exists(workspacePath)
+            ? Path.Combine(workspacePath, activeDocumentPath)
+            : workspacePath;
+
+        await WriteActiveDocumentToPath(workspacePack, fullSavePath, activeDocumentPath);
+    }
+
+    public async Task SaveActiveDocumentAsAsync(LoadedMarkerPack workspacePack, string activeDocumentPath)
+    {
+        var dialog = new SaveFileDialog
+        {
+            Title = "Save Marker File As",
+            Filter = "XML Marker File (*.xml)|*.xml",
+            FileName = activeDocumentPath.StartsWith("Untitled") ? "MyRoute.xml" : Path.GetFileName(activeDocumentPath)
+        };
+
+        if (dialog.ShowDialog() != true) return;
+
+        var newFilePath = dialog.FileName;
+
+        if (Path.GetExtension(newFilePath).ToLowerInvariant() is ".taco" or ".zip")
+        {
+            _dialogService.ShowError("Invalid Save Location", "Saving directly into a .taco or .zip archive is not supported. Please save as a standard .xml file.");
+            return;
+        }
+
+        var markersToSave = workspacePack.MarkersByFile[activeDocumentPath];
+        workspacePack.UnmanagedPois.TryGetValue(activeDocumentPath, out var unmanagedElements);
+
+        var newDoc = new XDocument(new XDeclaration("1.0", "utf-8", null), new XElement(TacoXmlConstants.OverlayDataElement));
+        _packWriterService.RewritePoisSection(newDoc, markersToSave, unmanagedElements ?? Enumerable.Empty<XElement>());
+
+        if (!await WriteDocumentToDiskAsync(newDoc, newFilePath)) return;
+
+        _logger.LogInformation("File saved as '{newFilePath}' and workspace context has been reset to a single-file mode.", newFilePath);
+    }
+
+    private async Task WriteActiveDocumentToPath(LoadedMarkerPack workspacePack, string saveDiskPath, string sourceKeyForWorkspace)
+    {
+        if (!workspacePack.XmlDocuments.TryGetValue(sourceKeyForWorkspace, out var docToSave))
+        {
+            _logger.LogError("Attempted to save '{sourceKey}' but its XDocument was not found.", sourceKeyForWorkspace);
+            _dialogService.ShowError("Save Error", "Could not find the active document's XML data to save.");
+            return;
+        }
+
+        workspacePack.UnmanagedPois.TryGetValue(sourceKeyForWorkspace, out var unmanagedElements);
+        var markersInOrder = workspacePack.MarkersByFile[sourceKeyForWorkspace];
+        _packWriterService.RewritePoisSection(docToSave, markersInOrder, unmanagedElements ?? Enumerable.Empty<XElement>());
+
+        if (!await WriteDocumentToDiskAsync(docToSave, saveDiskPath)) return;
+
+        workspacePack.AddedMarkers.RemoveWhere(m => m.SourceFile.Equals(sourceKeyForWorkspace, StringComparison.OrdinalIgnoreCase));
+        workspacePack.DeletedMarkers.RemoveWhere(m => m.SourceFile.Equals(sourceKeyForWorkspace, StringComparison.OrdinalIgnoreCase));
+        foreach (var marker in markersInOrder)
+        {
+            marker.IsDirty = false;
+        }
+
+        _logger.LogInformation("Document {docKey} saved successfully to {path}", sourceKeyForWorkspace, saveDiskPath);
+    }
+
+    private async Task<bool> WriteDocumentToDiskAsync(XDocument doc, string path)
+    {
+        try
+        {
+            var tempPath = Path.GetTempFileName();
+            await using (var writer = File.CreateText(tempPath))
+            {
+                await doc.SaveAsync(writer, SaveOptions.None, CancellationToken.None);
+            }
+            File.Move(tempPath, path, true);
+            _logger.LogInformation("Successfully wrote document to disk at {Path}", path);
+            return true;
+        }
+        catch (Exception ex)
+        {
+            _logger.LogError(ex, "Failed to write document to disk at {Path}", path);
+            _dialogService.ShowError("Save Failed", $"An error occurred while writing the file to disk:\n\n{ex.Message}");
+            return false;
+        }
+    }
+
+    private void ShowLoadPackErrors(PackLoadResult result)
+    {
+        var errorReport = new StringBuilder();
+        errorReport.AppendLine("The pack loaded, but some files could not be read. This is usually due to data corruption in the pack file itself.\n");
+        errorReport.AppendLine("Failed files:");
+        foreach (var error in result.Errors)
+        {
+            errorReport.AppendLine($" - {error.FileName}: {error.ErrorMessage}");
+            _logger.LogWarning("Load error in {FileName}: {ErrorMessage}", error.FileName, error.ErrorMessage);
+        }
+        _dialogService.ShowError("Pack Loaded with Errors", errorReport.ToString());
+    }
+}
